Rename html-to-mfm to fromHtml
